this commit has everything of softmax-prop impl. all tests pass on GPU but cpu/gpu discrepancy exists within softmax_backward_data when sharding on dim -1
