+++ spmd/tensor/ops/math_ops.py
@@ -11,7 +11,6 @@ from spmd.tensor.placement_types import (
 )
 from spmd.tensor.ops.utils import as_list, register_prop_rule
 
-
 def _gen_spec_with_pending_sum(
     spec: DTensorSpec, pending_sum: List[int]
 ) -> DTensorSpec:
@@ -224,7 +223,7 @@ def softmax_rule(op_schema: OpSchema) -> OutputSharding:
     dim_map = input_spec.dim_map
     softmax_dim = cast(
         int, op_schema.args_schema[len(op_schema.args_spec)]
-    )  # Is it better to put it into kwargs? e.g. op_schema.kwargs_schema['dim']
+    )
     output_spec: OutputSpecType = op_schema.args_spec[0]
     schema_suggestion = None
     failed_reason = None
@@ -248,3 +247,21 @@ def softmax_rule(op_schema: OpSchema) -> OutputSharding:
             "Cannot run _softmax on batch dim, need to replicate the tensor."
         )
     return OutputSharding(output_spec, schema_suggestion, failed_reason)
+
+
+@register_prop_rule("aten._softmax_backward_data.default")
+def softmax_bwd_rule(op_schema: OpSchema) -> OutputSharding:
+    grad_out_spec, out_spec, softmax_dim, _ = op_schema.args_schema
+    ops_dim_map = [spec.dim_map for spec in [grad_out_spec, out_spec]]
+    ops_dim_map = list(zip(*ops_dim_map))
+    if softmax_dim < len(ops_dim_map) and 0 in ops_dim_map[softmax_dim]:
+        schema_suggestion = OpSchema(
+            tuple([DTensorSpec(spec.mesh, [Replicate()], spec.shape, ndim=spec.ndim) for spec in input_specs]),
+            {},
+        )
+        failed_reason = "Cannot run _softmax_backward_data on batch dim, need to replicate the tensor."
+        _inplace_rewrap_schema_suggestion(schema_suggestion, op_schema)
+        return OutputSharding(None, [schema_suggestion], failed_reason)
+    else:
+        from spmd.tensor.ops.pointwise_ops import pointwise_rule
+        return pointwise_rule(op_schema)
